move file

+++ src/components/Modals/DeleteFiles.jsx
@@ -1,28 +1,57 @@
 import React, { useState, useContext } from "react";
 import { Button, Modal } from "reactstrap";
 import { FileOperationsContext } from "../../contexts/FileOperationsContext";
+import { FileContext } from "../../contexts/FileListContext";
 
 const DeleteFiles = ({ buttonLabel, buttonIcon, modalClassName }) => {
   const { deleteFileAndFolder, isDeletePending, isDeleteError } = useContext(
     FileOperationsContext
   );
-  const [modal, setModal] = useState(false);
+  const { getSelectedFiles } = useContext(FileContext);
+  const [isAbleToDelete, setIsAbleToDelete] = useState(false);
+  const isDeletingNonEmptyFolder = !!getSelectedFiles().find(
+    (item) => item.kind === "FOLDER"
+  );
+  const [showDeleteModal, setShowDeleteModal] = useState(false);
+  const [showCorfirmInputModal, setShowConfirmInputModal] = useState(false);
 
-  const toggle = () => {
-    setModal(!modal);
-  };
+  function toggleConfirmDeleteModal() {
+    setShowDeleteModal(!showDeleteModal);
+  }
 
-  const onHandleDelete = () => {
-    deleteFileAndFolder();
-  };
+  function onHandleDelete() {
+    if (isDeletingNonEmptyFolder) {
+      setShowConfirmInputModal(true);
+    } else {
+      deleteFileAndFolder();
+    }
+  }
+
+  function handleDeleteNonEmptyFolder() {
+    if (isAbleToDelete) {
+      deleteFileAndFolder();
+      setShowConfirmInputModal(false);
+      setShowDeleteModal(false);
+    }
+  }
+
+  function hideConfirmInputModal() {
+    setShowConfirmInputModal(false);
+    setShowDeleteModal(false);
+  }
 
   return (
     <>
-      <Button color="link" type="button" onClick={toggle}>
+      <Button color="link" type="button" onClick={toggleConfirmDeleteModal}>
         {buttonIcon}
         <span className="btn-inner--text">{buttonLabel}</span>
       </Button>
-      <Modal className={modalClassName} isOpen={modal} toggle={toggle}>
+
+      <Modal
+        className={modalClassName}
+        isOpen={showDeleteModal && !showCorfirmInputModal}
+        toggle={toggleConfirmDeleteModal}
+      >
         <div className="modal-header">
           <h5 className="modal-title" id="modal-title-default">
             Delete
@@ -32,7 +61,7 @@ const DeleteFiles = ({ buttonLabel, buttonIcon, modalClassName }) => {
             className="close"
             data-dismiss="modal"
             type="button"
-            onClick={toggle}
+            onClick={toggleConfirmDeleteModal}
           >
             <span aria-hidden={true}>×</span>
           </button>
@@ -41,14 +70,16 @@ const DeleteFiles = ({ buttonLabel, buttonIcon, modalClassName }) => {
           <p>Deleted items are removed from Qubee permanently.</p>
           <p>Do you wish to proceed?</p>
           {isDeleteError && (
-            <div className="alert alert-danger">
-              {isDeleteError.message}
-            </div>
+            <div className="alert alert-danger">{isDeleteError.message}</div>
           )}
         </div>
         <div className="modal-footer">
           <div className="text-center">
-            <Button data-dismiss="modal" color="link" onClick={toggle}>
+            <Button
+              data-dismiss="modal"
+              color="link"
+              onClick={toggleConfirmDeleteModal}
+            >
               Cancel
             </Button>
             <Button
@@ -56,6 +87,60 @@ const DeleteFiles = ({ buttonLabel, buttonIcon, modalClassName }) => {
               type="submit"
               onClick={onHandleDelete}
               disabled={isDeletePending}
+            >
+              {isDeletePending ? "Please wait" : "Delete"}
+            </Button>
+          </div>
+        </div>
+      </Modal>
+
+      <Modal
+        className={modalClassName}
+        isOpen={showCorfirmInputModal}
+        toggle={hideConfirmInputModal}
+      >
+        <div className="modal-header">
+          <h5 className="modal-title" id="modal-title-default">
+            Confirm delete
+          </h5>
+          <button
+            aria-label="Close"
+            className="close"
+            data-dismiss="modal"
+            type="button"
+            onClick={hideConfirmInputModal}
+          >
+            <span aria-hidden={true}>×</span>
+          </button>
+        </div>
+        <div className="modal-body" style={{ textAlign: "center" }}>
+          <p>The folder you are trying to delete is not empty</p>
+          <p>
+            Please type <strong>DELETE</strong> to confirm your action
+          </p>
+          <input
+            type="text"
+            onChange={(e) =>
+              setIsAbleToDelete(
+                e.target.value && e.target.value.toLowerCase() === "delete"
+              )
+            }
+          />
+        </div>
+        <div className="modal-footer">
+          <div className="text-center">
+            <Button
+              data-dismiss="modal"
+              color="link"
+              onClick={hideConfirmInputModal}
+            >
+              Cancel
+            </Button>
+            <Button
+              color="primary"
+              type="submit"
+              onClick={handleDeleteNonEmptyFolder}
+              disabled={isDeletePending || !isAbleToDelete}
             >
               {isDeletePending ? "Please wait" : "Delete"}
             </Button>
